+++ src/main/java/Dienblad.java
@@ -51,11 +51,11 @@ public class Dienblad {
      */
     public double getTotaalPrijs() {
         // method body omitted
-        int totaal = 0;
+        double totaal = 0;
         Iterator<Artikel> it = artikelen.iterator();
         while(it.hasNext()) {
             Artikel artikel = it.next();
-            totaal += artikel.getPrijs().intValue();
+            totaal += artikel.getPrijs();
         }
         return totaal;
     }
